Add CustomStartPositionMediaSource to enable use of initialTime parameter for a track on Android.
This method provides a simple wrapper for a media source that allows setting a start position.

This retains the ability to seek back before the initial time, unlike using cropped media source.

It also avoid relying on the onPositionDiscontinuity callback, which sometimes allows a short period of the beginning of the track to be heard.